Dog Bone update
Not final, I have an explosion object prefab, which I would ideally use, and a separate particle system attached as a component of the bone object. The prefab is in the bone object, but acts as a separate object. Neither will destroy, although I have an explosion and destruction script (called Ammo.cs) and it is listed as a component of the bone object. I hope to finish it this afternoon (Friday, March 2nd)